Require a URI directly to the catalog
Previously, we required a CatalogRootUri, which was a URI to the catalog
s containing directory (or however that mapped to the backend we
determined from the URI). From there, we assumed that the catalog's full
URI was CatalogRootUri/BoxName.json.

This required passing both root URI and box name when querying the
catalog, which didn't seem right. It was also potentially confusing
for users.

With this change, only require a box name for adding boxes, determining
everything else by the direct catalog URI itself.